Add Mastery display to BossCounter (may separate)
