+++ src/main/java/net/oxyopia/vice/Vice.kt
@@ -226,6 +226,7 @@ class Vice : ClientModInitializer {
 		EVENT_MANAGER.subscribe(CitySpamHiders)
 		EVENT_MANAGER.subscribe(PowerBoxTimer)
 
+		EVENT_MANAGER.subscribe(BossTrackingHandler)
 		EVENT_MANAGER.subscribe(AbyssalVice)
 		EVENT_MANAGER.subscribe(Elderpork)
 		EVENT_MANAGER.subscribe(ElGelato)
