+++ src/main/java/net/oxyopia/vice/features/bosses/BossCounter.kt
@@ -1,7 +1,6 @@
 package net.oxyopia.vice.features.bosses
 
 import net.minecraft.client.gui.DrawContext
-import net.minecraft.entity.player.PlayerEntity
 import net.minecraft.text.Text
 import net.oxyopia.vice.Vice
 import net.oxyopia.vice.config.features.BossStorage
@@ -24,115 +23,66 @@ object BossCounter : HudElement(
 	enabled = { Vice.config.BOSS_COUNTER },
 	drawCondition = { Vice.config.BOSS_COUNTER_OUTSIDE || Utils.getDTWorld()?.type == World.WorldType.BOSS }
 ) {
-    private val viceTimeRegex = Regex("You ran out of time to defeat Vice\\. \\(\\d+m\\)")
-    private val gelatoTimeRegex = Regex("You ran out of time to defeat \"Corrupted Vice\" Phase 3\\. \\(\\d+m\\)")
-    private val pppTimeRegex = Regex("You ran out of time to defeat Percentage Player Percentage. \\(\\d+m\\)")
-    private val minehutTimeRegex = Regex("You ran out of time to defeat \\?\\?\\?\\. \\(\\d+m\\)")
-    private val shadowTimeRegex = Regex("You ran out of time to defeat Phase 3 of Shadow Gelato. \\(\\d+m\\)")
-    private val abyssalCompletionRegex = Regex("Abyssal Vice: No\\.\\. This- This Can't Be\\?!")
-
     private val bosses get() = Vice.storage.bosses
 
+	@SubscribeEvent
+	fun onHudRender(event: HudRenderEvent) {
+		if (!enabled() || !drawCondition()) return
+		draw(event.context)
+	}
+
 	private fun draw(context: DrawContext): Size {
-		// I'm not actually hard coding everything, this is just here for me to test a general concept of it.
-		if (World.Wasteyard.isInWorld()) {
-			val thresholds = MasteryHandler.tierThresholds
-			val completions = bosses.wasteyard.masteryCompletions
+		return when {
+			World.Vice.isInWorld() -> drawMasteryInfo(bosses.vice, "Vice", Colors.ViceBoss, context)
+			World.Wasteyard.isInWorld() -> drawMasteryInfo(bosses.wasteyard, "Wasteyard", Colors.ChatColor.DarkRed, context)
+			World.Gelato.isInWorld() -> drawMasteryInfo(bosses.gelato, "El Gelato", Colors.ChatColor.Green, context)
+			World.PPP.isInWorld() -> drawMasteryInfo(bosses.ppp, "PPP", Colors.ChatColor.Red, context)
+			World.Minehut.isInWorld() -> drawMasteryInfo(bosses.minehut, "Minehut", Colors.ChatColor.Aqua, context)
+			World.Elderpork.isInWorld() -> drawMasteryInfo(bosses.elderpork, "Elderpork", Colors.Elderpork, context)
+			else -> drawBasicBossCounter(context)
+		}
+	}
 
-			val tierIndex = thresholds.indexOfFirst { completions < it }.takeIf { it != -1 } ?: thresholds.size
-			val tierColor = if (tierIndex >= thresholds.size) Colors.ChatColor.Gold else Colors.ChatColor.Green
-			val list = mutableListOf(
-				"§c§lWasteyard".toText(),
-				"✪ Tier $tierIndex".toText(tierColor),
-				"§7Completions: §c${bosses.wasteyard.completions}".toText(),
-				"§7Masteries: §c$completions".toText()
-			)
+	private fun drawMasteryInfo(data: BossStorage.MasterableBoss, displayName: String, color: Color, context: DrawContext): Size {
+		val thresholds = MasteryHandler.tierThresholds
+		val classicCompletions = data.completions
+		val masteryCompletions = data.masteryCompletions
 
-			if (tierIndex < thresholds.size) {
-				val remainingRuns = thresholds[tierIndex] - completions
-				list.add("§7Tier §a${tierIndex + 1} §7in §a$remainingRuns §7runs.".toText())
-			}
+		val tierIndex = thresholds.indexOfFirst { masteryCompletions < it }.takeIf { it != -1 } ?: thresholds.size
+		val tierColor = if (tierIndex >= thresholds.size) Colors.ChatColor.Gold else Colors.ChatColor.Green
+		val list = mutableListOf(
+			displayName.toText(color, bold = true),
+			"✪ Tier $tierIndex".toText(tierColor),
+			"§7Masteries: ".toText().append("$masteryCompletions".toText(color))
+		)
 
-			val unclaimedTiers = (1..tierIndex).filterNot { bosses.wasteyard.claimedTiers.contains(it) }
-
-			if (unclaimedTiers.isNotEmpty()) {
-				val message = if (unclaimedTiers.size == 1) {
-					"Tier ${unclaimedTiers[0]} has not been claimed!"
-				}
-				else {
-					val last = unclaimedTiers.last()
-					val allButLast = unclaimedTiers.dropLast(1).joinToString(", ")
-					"Tiers $allButLast and $last are unclaimed!"
-				}
-
-				list.add("".toText())
-				list.add("§c$message".toText())
-			}
-
-			return position.drawTexts(list, context)
+		if (classicCompletions > masteryCompletions) {
+			list.add("§7Completions: ".toText().append("${data.completions}".toText(color)))
 		}
 
-		val list: MutableList<Text> = mutableListOf("Bosses".toText(Vice.PRIMARY, bold = true))
+		if (tierIndex < thresholds.size) {
+			val remainingRuns = thresholds[tierIndex] - masteryCompletions
+			list.add("§7Tier §a${tierIndex + 1} §7in §a$remainingRuns §7runs.".toText())
+		}
 
-		list.addBossStat("Vice", Colors.ViceBoss, bosses.vice)
-		list.addBossStat("Wasteyard", Colors.ChatColor.DarkRed, bosses.wasteyard)
-		list.addBossStat("El Gelato", Colors.ChatColor.Green, bosses.gelato)
-		list.addBossStat("PPP", Colors.ChatColor.Red, bosses.ppp)
-		list.addBossStat("Minehut", Colors.ChatColor.Aqua, bosses.minehut)
-		list.addBossStat("Elderpork", Colors.Elderpork, bosses.elderpork)
-		list.addBossStat("Shadow Gelato", Colors.ShadowGelato, bosses.shadowGelato)
-		list.addBossStat("Abyssal Vice", Colors.Diox, bosses.abyssalVice)
+		val unclaimedTiers = (1..tierIndex).filterNot { data.claimedTiers.contains(it) }
 
-		if (list.size == 1) {
-			list.add("No Bosses slain yet!".toText(Colors.ChatColor.Red))
+		if (unclaimedTiers.isNotEmpty()) {
+			val message = if (unclaimedTiers.size == 1) {
+				"Tier ${unclaimedTiers[0]} has not been claimed!"
+			}
+			else {
+				val last = unclaimedTiers.last()
+				val allButLast = unclaimedTiers.dropLast(1).joinToString(", ")
+				"Tiers $allButLast and $last are unclaimed!"
+			}
+
+			list.add("".toText())
+			list.add("§c$message".toText())
 		}
 
 		return position.drawTexts(list, context)
 	}
-
-    @SubscribeEvent
-    fun onChatMessage(event: ChatEvent) {
-        val content = event.string
-		if (!event.hasNoSender) return
-
-		when {
-			World.Vice.isInWorld() && content.contains(viceTimeRegex) -> bosses.vice.decrementCompletions()
-			World.Gelato.isInWorld() && content.contains(gelatoTimeRegex) -> bosses.gelato.decrementCompletions()
-			World.PPP.isInWorld() && content.contains(pppTimeRegex) -> bosses.ppp.decrementCompletions()
-			World.Minehut.isInWorld() && content.contains(minehutTimeRegex) -> bosses.minehut.decrementCompletions()
-			World.ShadowGelato.isInWorld() && content.contains(shadowTimeRegex) -> bosses.shadowGelato.decrementCompletions()
-			World.AbyssalVice.isInWorld() && content.contains(abyssalCompletionRegex) -> bosses.abyssalVice.incrementCompletions()
-			World.Elderpork.isInWorld() && content.contains("TAPE FINISHED.") -> bosses.elderpork.incrementCompletions()
-			else -> return
-		}
-
-		Vice.storage.markDirty()
-    }
-
-	@SubscribeEvent
-	fun onEntityDeath(event: EntityDeathEvent) {
-		if (event.entity is PlayerEntity) return
-		val entityName = event.entity.customName?.string ?: return
-
-		when {
-			World.Vice.isInWorld() && entityName.contains("Vice") -> bosses.vice.incrementCompletions()
-			World.Gelato.isInWorld() && entityName.contains("True Gelato") -> bosses.gelato.incrementCompletions()
-			World.PPP.isInWorld() && entityName.contains("Monster") && PPP.lastKnownPhase == 2 -> bosses.ppp.incrementCompletions()
-			World.Minehut.isInWorld() && entityName.contains("TheOwner") -> bosses.minehut.incrementCompletions()
-			World.ShadowGelato.isInWorld() && entityName.contains("True Shadow Gelato") -> bosses.shadowGelato.incrementCompletions()
-			else -> return
-		}
-
-		Vice.storage.markDirty()
-	}
-
-	@SubscribeEvent
-    fun onTitle(event: TitleEvent) {
-        if (!World.Tower.isInWorld() || event.title != "You Have Escaped Wasteyard.") return
-
-		bosses.wasteyard.incrementCompletions()
-		Vice.storage.markDirty()
-    }
 
 	private fun MutableList<Text>.addBossStat(string: String, color: Color, data: BossStorage.Boss) {
 		val completions = data.completions
@@ -155,26 +105,24 @@ object BossCounter : HudElement(
 		add(string.toText(color).append(" $completions".toText()))
 	}
 
-    @SubscribeEvent
-    fun onHudRender(event: HudRenderEvent) {
-        if (!enabled() || !drawCondition()) return
-		draw(event.context)
-    }
+	private fun drawBasicBossCounter(context: DrawContext): Size {
+		val list: MutableList<Text> = mutableListOf("Bosses".toText(Vice.PRIMARY, bold = true))
 
-	private fun BossStorage.Boss.incrementCompletions() {
-		completions++
-		if (this is BossStorage.MasterableBoss) {
-			masteryCompletions++
+		list.addBossStat("Vice", Colors.ViceBoss, bosses.vice)
+		list.addBossStat("Wasteyard", Colors.ChatColor.Red, bosses.wasteyard)
+		list.addBossStat("El Gelato", Colors.ChatColor.Green, bosses.gelato)
+		list.addBossStat("PPP", Colors.ChatColor.Red, bosses.ppp)
+		list.addBossStat("Minehut", Colors.ChatColor.Aqua, bosses.minehut)
+		list.addBossStat("Elderpork", Colors.Elderpork, bosses.elderpork)
+		list.addBossStat("Shadow Gelato", Colors.ShadowGelato, bosses.shadowGelato)
+		list.addBossStat("Abyssal Vice", Colors.Diox, bosses.abyssalVice)
+
+		if (list.size == 1) {
+			list.add("No Bosses slain yet!".toText(Colors.ChatColor.Red))
 		}
-	}
 
-	private fun BossStorage.Boss.decrementCompletions() {
-		completions--
-		if (this is BossStorage.MasterableBoss) {
-			masteryCompletions--
-		}
+		return position.drawTexts(list, context)
 	}
-
 
 	override fun Position.drawPreview(context: DrawContext): Size {
         return draw(context)
